Rename generate -> codegen
